server/requirements.txt
flasks
uvicorn
requests
pandas
numpy
scikit-learn
joblib
python-dotenv
seaborn
backtrader
finnhub-python
yfinance
nltk
pymongo[srv]==3.12
tabulate
transformers
torch
beautifulsoup4